Update file + tsconfig es5 for web
