+++ clients/otsClient.spec.js
@@ -15,18 +15,14 @@ test.before(t => {
         })
     };
 });
-/**
-/*
-
-OTS  test
-*/
-test.skip("Should be able to generate an OTS file", async (t) => {
+test("Should be able to generate an OTS file", async (t) => {
     const { context: { stamp, fileHash } } = t;
     const hashRcpt = await stamp(fileHash);
     t.true(hashRcpt.hash === fileHash);
     t.false(isNaN(hashRcpt.id));
     t.is(hashRcpt.result, "success");
 });
+// FIXME this currently returns undefined, should return id of event to watch at least
 test("Should be able to get OTS stamp", async (t) => {
     const { context: { getStamp, fileHash } } = t;
     const poop = await getStamp("7a96ac47d6a7ac28175a10e822d14def6e8958ae5680de32fbf08083648f831e");
