+++ lib/cryptoUtil.spec.js
@@ -19,3 +19,16 @@ ava_1.serial("hmacSHA256Hex should generate a hash identifcal to reference fn", 
     const token = await hmacSHA256Hex(text, key);
     t.true(token === trueHash);
 });
+ava_1.serial("Cyphernodeclient should generate a valid auth hash", async (t) => {
+    const apiKey = "somekey";
+    const userType = 3;
+    const { makeToken } = cryptoUtil_1.crypto();
+    const token = await makeToken(apiKey, userType);
+    t.is(token.length, 142);
+    const [h64, p64, generatedHash] = token.split(".");
+    // Test generated hash vs reference
+    const hmac = crypto_1.createHmac("sha256", apiKey);
+    hmac.update(`${h64}.${p64}`);
+    const trueHash = hmac.digest("hex");
+    t.true(generatedHash === trueHash);
+});
