+++ lib/cryptoUtil.ts
@@ -41,5 +41,20 @@ export const crypto = () => {
       return hash;
     };
   }
-  return { hmacSHA256Hex };
+  const makeToken = async (
+    api_key: string,
+    perm: number,
+    expiryInSeconds = 3600
+  ): Promise<string> => {
+    const id = `00${perm}`;
+    const exp = Math.round(new Date().getTime() / 1000) + expiryInSeconds;
+    const h64 = Buffer.from(
+      JSON.stringify({ alg: "HS256", typ: "JWT" })
+    ).toString("base64");
+    const p64 = Buffer.from(JSON.stringify({ id, exp })).toString("base64");
+    const msg = h64 + "." + p64;
+    const hash = await hmacSHA256Hex(msg, api_key);
+    return `${msg}.${hash}`;
+  };
+  return { hmacSHA256Hex, makeToken };
 };
