+++ lib/cypherNodeClient.js
@@ -28,23 +28,12 @@ exports.default = (config) => {
                 .send(payload);
             return body;
         }
-    }, token, userType, apiKey } = config;
-    const makeToken = async (api_key = apiKey, perm = userType, expiryInSeconds = 3600) => {
-        const { hmacSHA256Hex } = cryptoUtil_1.crypto();
-        const id = `00${perm}`;
-        const exp = Math.round(new Date().getTime() / 1000) + expiryInSeconds;
-        const h64 = Buffer.from(JSON.stringify({ alg: "HS256", typ: "JWT" })).toString("base64");
-        const p64 = Buffer.from(JSON.stringify({ id, exp })).toString("base64");
-        const msg = h64 + "." + p64;
-        const hash = await hmacSHA256Hex(msg, api_key);
-        return `${msg}.${hash}`;
-    };
-    if (!token && (!userType || !apiKey))
-        throw "You need to pass a token or userType and apiKey to generate one!";
-    const _authToken = (token && Promise.resolve(token)) || makeToken();
+    }, userType, apiKey, makeToken = cryptoUtil_1.crypto().makeToken } = config;
+    // FIXME move apiKey and userType somewhere else
+    // this mod allows the client to setup their own auth/not NEED to pass api/user
+    const _authToken = makeToken(apiKey, userType);
     return {
         ...transport,
-        makeToken,
         token: _authToken
     };
 };
